Name Identification for enemies on AbstractHippyRobot
Removed HippyName class as it is not necessary.
