+++ robocode.ui/src/main/java/net/sf/robocode/ui/battleview/BattleView.java
@@ -40,7 +40,7 @@ import static java.lang.Math.*;
 
 import java.util.Random;
 
-import foobar.hippy.HippyName;
+import foobar.hippy.AbstractHippyRobot;
 
 /**
  * @author Mathew A. Nelson (original)
@@ -417,7 +417,7 @@ public class BattleView extends Canvas {
 		if (drawScanArcs) {
 			for (IRobotSnapshot robotSnapshot : snapShot.getRobots()) {
 				if (robotSnapshot.getState().isAlive()) {
-					if (HippyName.isVan(robotSnapshot.getName())) {
+					if (AbstractHippyRobot.isVan(robotSnapshot.getName())) {
 						drawScanArc(g, robotSnapshot);
 					}
 				}
@@ -501,7 +501,7 @@ public class BattleView extends Canvas {
 					at = AffineTransform.getTranslateInstance(x, y);
 					at.rotate(robotSnapshot.getRadarHeading());
 
-					if (HippyName.isVan(robotSnapshot.getName())) {
+					if (AbstractHippyRobot.isVan(robotSnapshot.getName())) {
 						RenderImage radarRenderImage = imageManager
 								.getColoredRadarRenderImage(robotSnapshot
 										.getRadarColor());
